+++ docs/trace/reporting-exceptions/Program.cs
@@ -19,6 +19,8 @@ using System.Diagnostics;
 using OpenTelemetry;
 using OpenTelemetry.Trace;
 
+namespace ReportingExceptions;
+
 public class Program
 {
     private static readonly ActivitySource MyActivitySource = new(